+++ examples/netcore22/XPikeConfig22/Startup.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using XPike.IoC.SimpleInjector.AspNetCore;
 
 namespace XPikeConfig22
 {
@@ -32,11 +33,15 @@ namespace XPikeConfig22
 
 
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+
+            services.AddXPikeDependencyInjection();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env)
         {
+            app.UseXPikeDependencyInjection();
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();