+++ examples/netcore22/XPikeConfig22MS/Models/ErrorViewModel.cs
@@ -1,0 +1,9 @@
+namespace XPikeConfig22MS.Models
+{
+    public class ErrorViewModel
+    {
+        public string RequestId { get; set; }
+
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    }
+}